Rename constructor function builder file to match canonical declaration
