+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/constructor-function/buildConstructorFunction.ts
