+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder.ts
@@ -1,4 +1,4 @@
-import { GenericPropertyNameTuple } from './constructorFunctionBuilder';
+import { GenericPropertyNameTuple } from './buildConstructorFunction';
 import {
   NamedConstructorFunctionInstanceTypeBuilderParent,
   buildNamedConstructorFunctionInstanceTypeBuilder,
