+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/constructor-function/namedConstructorFunctionInstanceTypeBuilder.ts
@@ -6,7 +6,7 @@ import {
   GenericConstructedInstance,
   ConstructorInputTransformer,
   GenericConstructorInputTransformer,
-} from './constructorFunctionBuilder';
+} from './buildConstructorFunction';
 import {
   NamedConstructorFunctionAssemblerContext,
   NamedConstructorFunctionAssemblerParent,
